Refactor constants, utilities and types into lib
