+++ apps/site/.env.local.example
@@ -1,8 +1,8 @@
 # API key for SimpleHash
-NEXT_PUBLIC_API_KEY=
+API_KEY=
 
 # API key for CryptoCompare
-NEXT_PUBLIC_CC_API_KEY=
+CC_API_KEY=
 
 # hostname
 NEXT_PUBLIC_BASE_URL=localhost:4200
