+++ apps/site/src/components/PageLanding/Featured.tsx
@@ -1,10 +1,8 @@
-import { NFT } from '@myflownfts/data-access';
+import { NFT } from '@data-access';
 import useTranslation from 'next-translate/useTranslation';
 import { useRef } from 'react';
-import { NFT } from '../../api/types';
-import { ScrollButtons } from '../common';
+import { NFTCard, ScrollButtons } from '../common';
 import { ReactComponent as StarIcon } from '../common/images/icon-section-star.svg';
-import NFTCard from '../common/NFTCard';
 
 interface Props {
   nfts: NFT[];
