+++ apps/site/src/components/PageLanding/Header.tsx
@@ -1,3 +1,5 @@
+import { addressAtom, isLandingPageLoadedAtom, userAtom } from '@web/atoms';
+import { useWallet } from '@web/utils';
 import classNames from 'classnames';
 import { useAtom } from 'jotai';
 import useTranslation from 'next-translate/useTranslation';
@@ -5,8 +7,6 @@ import Image from 'next/future/image';
 import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react';
 import { toast } from 'react-toastify';
-import { addressAtom, isLandingPageLoadedAtom, userAtom } from '../../atoms';
-import { useWallet } from '../../utils';
 import { Button, TextInput } from '../common';
 import { ReactComponent as SearchIcon } from '../common/images/icon-search.svg';
 import bgHeaderMobileSm from './images/header/bg-header-mobile-sm.png';
