+++ apps/site/src/components/PageViewNFT/index.tsx
@@ -1,4 +1,17 @@
+import { BASE_URL, ExchangeRates, NFT } from '@data-access';
 import * as Popover from '@radix-ui/react-popover';
+import {
+  getExchangeRates,
+  getNFTByTokenId,
+  postRefreshMetadata,
+} from '@web/api';
+import { ReactComponent as CloseIcon } from '@web/components/common/images/icon-close.svg';
+import { ReactComponent as EthIcon } from '@web/components/common/images/icon-eth.svg';
+import { ReactComponent as FlowIcon } from '@web/components/common/images/icon-flow-nofill.svg';
+import { ReactComponent as LinkIcon } from '@web/components/common/images/icon-link.svg';
+import { ReactComponent as RefreshIcon } from '@web/components/common/images/icon-refresh.svg';
+import { ReactComponent as ShareIcon } from '@web/components/common/images/icon-share.svg';
+import { ReactComponent as USDIcon } from '@web/components/common/images/icon-usd.svg';
 import classNames from 'classnames';
 import { NextSeo } from 'next-seo';
 import useTranslation from 'next-translate/useTranslation';
@@ -15,22 +28,6 @@ import {
 } from 'react-share';
 import { toast } from 'react-toastify';
 import ReactTooltip from 'react-tooltip';
-
-import {
-  ExchangeRates,
-  getExchangeRates,
-  getNFTByTokenId,
-  NFT,
-  postRefreshMetadata,
-} from '@myflownfts/data-access';
-import { ReactComponent as CloseIcon } from '../../components/common/images/icon-close.svg';
-import { ReactComponent as EthIcon } from '../../components/common/images/icon-eth.svg';
-import { ReactComponent as FlowIcon } from '../../components/common/images/icon-flow-nofill.svg';
-import { ReactComponent as LinkIcon } from '../../components/common/images/icon-link.svg';
-import { ReactComponent as RefreshIcon } from '../../components/common/images/icon-refresh.svg';
-import { ReactComponent as ShareIcon } from '../../components/common/images/icon-share.svg';
-import { ReactComponent as USDIcon } from '../../components/common/images/icon-usd.svg';
-import { BASE_URL } from '../../constants';
 
 import { Button, Chip, Footer, Loader, Navbar } from '../common';
