--- apps/site/src/constants.ts
@@ -1,36 +1,0 @@
-export const BASE_API_URL = 'https://api.simplehash.com/api/v0';
-export const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL ?? '';
-export const CURRENCY_API_URL = 'https://min-api.cryptocompare.com/data';
-
-export const ACCESS_NODE = 'https://rest-mainnet.onflow.org';
-export const DISCOVERY_WALLET = 'https://flow-wallet.blocto.app/authn';
-export const PROFILE_CONTRACT = '0x97bafa4e0b48eef';
-
-export const DEFAULT_FEATURED_NFTS = [
-  'flow.A.0b2a3299cc857e29.TopShot.37143972',
-  'flow.A.329feb3ab062d289.UFC_NFT.1585361',
-  'flow.A.921ea449dffec68a.Flovatar.33',
-  'flow.A.807c3d470888cc48.Flunks.9448',
-  'flow.A.5b82f21c0edf76e3.StarlyCard.1668',
-  'flow.A.f2af175e411dfff8.MetaPanda.864',
-  'flow.A.2068315349bdfce5.GoatedGoats.5521',
-  'flow.A.28abb9f291cadaf2.BarterYardClubWerewolf.3783',
-  'flow.A.7c8995e83c4b1843.Collector.1',
-  'flow.A.9d21537544d9123d.Momentables.1083',
-  'flow.A.329feb3ab062d289.RareRooms_NFT.51710',
-  'flow.A.329feb3ab062d289.Canes_Vault_NFT.1902509',
-  'flow.A.329feb3ab062d289.RaceDay_NFT.1400982',
-  'flow.A.329feb3ab062d289.The_Next_Cartel_NFT.32565',
-  'flow.A.8529aaf64c168952.MonoCatMysteryBox.41',
-  'flow.A.329feb3ab062d289.UFC_NFT.1733389',
-  'flow.A.0b2a3299cc857e29.TopShot.650',
-  'flow.A.8b148183c28ff88f.Gaia.589',
-  'flow.A.807c3d470888cc48.Flunks.0',
-  'flow.A.f2af175e411dfff8.MetaPanda.13',
-  'flow.A.7c8995e83c4b1843.Collector.28',
-  'flow.A.799da0ef17f38104.DaysOnFlow.30',
-  'flow.A.758252ab932a3416.YahooCollectible.4',
-  'flow.A.28abb9f291cadaf2.BarterYardClubWerewolf.20',
-  'flow.A.807c3d470888cc48.Backpack.48',
-  'flow.A.329feb3ab062d289.RaceDay_NFT.7111',
-];
