+++ apps/site/src/pages/nft/[contract_address]/[token_id].tsx
@@ -1,3 +1,3 @@
-import PageViewNFT from '../../../components/PageViewNFT';
+import PageViewNFT from '@web/components/PageViewNFT';
 
 export default PageViewNFT;
