+++ apps/site/src/pages/owned/[address].tsx
@@ -1,3 +1,3 @@
-import PageViewNFTs from '../../components/PageViewNFTs';
+import PageViewNFTs from '@web/components/PageViewNFTs';
 
 export default PageViewNFTs;
