+++ libs/data-access/src/index.ts
@@ -1,2 +1,3 @@
-export * from './lib/data-access';
+export * from './lib/constants';
 export * from './lib/types';
+export * from './lib/utilities';
