+++ libs/data-access/src/lib/utilities.ts
@@ -1,9 +1,10 @@
 export const fetchFromApi = async (url: string) => {
+  if (!process.env['API_KEY']) return;
   return fetch(url, {
     method: 'GET',
     headers: {
       Accept: 'application/json',
-      'X-API-KEY': process.env.API_KEY,
+      'X-API-KEY': process.env['API_KEY'],
     },
   });
 };
@@ -13,17 +14,18 @@ export const fetchFromCurrencyApi = async (url: string) => {
     method: 'GET',
     headers: {
       Accept: 'application/json',
-      authorization: `Apikey ${process.env.CURRENCY_API_KEY}`,
+      authorization: `Apikey ${process.env['CURRENCY_API_KEY']}`,
     },
   });
 };
 
 export const postToApi = async (url: string) => {
+  if (!process.env['API_KEY']) return;
   return fetch(url, {
     method: 'POST',
     headers: {
       Accept: 'application/json',
-      'X-API-KEY': process.env.API_KEY,
+      'X-API-KEY': process.env['API_KEY'],
     },
   });
 };
